+++ src/bika/lims/controlpanel/bika_attachmenttypes.py
@@ -35,93 +35,10 @@ from senaite.core.interfaces import IHideActionsMenu
 from zope.interface.declarations import implements
 
 
-# TODO: Separate content and view into own modules!
-
-
-class AttachmentTypesView(BikaListingView):
-
-    def __init__(self, context, request):
-        super(AttachmentTypesView, self).__init__(context, request)
-
-        self.catalog = "senaite_catalog_setup"
-
-        self.contentFilter = {
-            "portal_type": "AttachmentType",
-            "sort_on": "sortable_title",
-            "sort_order": "ascending",
-        }
-
-        self.context_actions = {
-            _("Add"): {
-                "url": "createObject?type_name=AttachmentType",
-                "permission": AddAttachmentType,
-                "icon": "++resource++bika.lims.images/add.png"}
-        }
-
-        self.title = self.context.translate(_("Attachment Types"))
-        self.icon = "{}/{}".format(
-            self.portal_url,
-            "/++resource++bika.lims.images/attachment_big.png"
-        )
-
-        self.show_select_row = False
-        self.show_select_column = True
-        self.pagesize = 25
-
-        self.columns = collections.OrderedDict((
-            ("Title", {
-                "title": _("Attachment Type"),
-                "index": "sortable_title"}),
-            ("Description", {
-                "title": _("Description"),
-                "index": "Description",
-                "toggle": True,
-            }),
-        ))
-
-        self.review_states = [
-            {
-                "id": "default",
-                "title": _("Active"),
-                "contentFilter": {"is_active": True},
-                "transitions": [{"id": "deactivate"}, ],
-                "columns": self.columns.keys(),
-            }, {
-                "id": "inactive",
-                "title": _("Inactive"),
-                "contentFilter": {'is_active': False},
-                "transitions": [{"id": "activate"}, ],
-                "columns": self.columns.keys(),
-            }, {
-                "id": "all",
-                "title": _("All"),
-                "contentFilter": {},
-                "columns": self.columns.keys(),
-            },
-        ]
-
-    def folderitem(self, obj, item, index):
-        """Service triggered each time an item is iterated in folderitems.
-        The use of this service prevents the extra-loops in child objects.
-        :obj: the instance of the class to be foldered
-        :item: dict containing the properties of the object to be used by
-            the template
-        :index: current index of the item
-        """
-        obj = api.get_object(obj)
-        title = obj.Title()
-        description = obj.Description()
-        url = obj.absolute_url()
-
-        item["replace"]["Title"] = get_link(url, value=title)
-        item["Description"] = description
-
-        return item
-
-
 schema = ATFolderSchema.copy()
 
 
+# TODO: Migrated to DX - https://github.com/senaite/senaite.core/pull/####
 class AttachmentTypes(ATFolder):
     implements(IAttachmentTypes, IHideActionsMenu)
     displayContentsTab = False
