+++ src/bika/lims/profiles/default/factorytool.xml
@@ -9,7 +9,6 @@
   <type portal_type="AnalysisService"/>
   <type portal_type="AnalysisSpec"/>
   <type portal_type="Attachment"/>
-  <type portal_type="AttachmentType"/>
   <type portal_type="Batch"/>
   <type portal_type="BatchLabel"/>
   <type portal_type="BikaSetup"/>
