+++ src/bika/lims/profiles/default/structure/bika_setup/.preserve
@@ -3,7 +3,6 @@ bika_analysiscategories
 bika_analysisservices
 bika_analysisspecs
 bika_artemplates
-bika_attachmenttypes
 bika_batchlabels
 bika_calculations
 bika_containertypes
