--- src/bika/lims/profiles/default/structure/bika_setup/bika_attachmenttypes/.properties
@@ -1,3 +1,0 @@
-[DEFAULT]
-description = 
-title = Attachment Types
