+++ src/bika/lims/profiles/default/types.xml
@@ -14,8 +14,6 @@
  <object name="AnalysisSpec" meta_type="Factory-based Type Information with dynamic views"/>
  <object name="AnalysisSpecs" meta_type="Factory-based Type Information with dynamic views"/>
  <object name="Attachment" meta_type="Factory-based Type Information with dynamic views"/>
- <object name="AttachmentType" meta_type="Factory-based Type Information with dynamic views"/>
- <object name="AttachmentTypes" meta_type="Factory-based Type Information with dynamic views"/>
  <object name="Batch" meta_type="Factory-based Type Information with dynamic views"/>
  <object name="BatchFolder" meta_type="Factory-based Type Information with dynamic views"/>
  <object name="BatchLabel" meta_type="Factory-based Type Information with dynamic views"/>
