--- src/bika/lims/profiles/default/types/AttachmentTypes.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<object name="AttachmentTypes"
-        meta_type="Factory-based Type Information with dynamic views"
-        xmlns:i18n="http://xml.zope.org/namespaces/i18n"
-        i18n:domain="senaite.core"
-        purge="True">
-
- <property name="title" i18n:translate="">Attachment Types</property>
- <property name="description"></property>
- <property name="content_icon">senaite_theme/icon/attachment</property>
- <property name="content_meta_type">AttachmentTypes</property>
- <property name="product">bika.lims</property>
- <property name="factory">addAttachmentTypes</property>
- <property name="add_view_expr"></property>
- <property name="link_target"></property>
- <property name="global_allow">False</property>
- <property name="filter_content_types">True</property>
- <property name="allowed_content_types">
-  <element value="AttachmentType"/>
- </property>
- <property name="allow_discussion">False</property>
- <property name="default_view_fallback">False</property>
-
- <alias from="(Default)" to="folder_view"/>
- <alias from="view" to="folder_view"/>
- <alias from="base_view" to="folder_view"/>
- <alias from="edit" to="base_edit"/>
-
-</object>
