+++ src/senaite/core/browser/controlpanel/configure.zcml
@@ -21,6 +21,7 @@
 
   <!-- Include controlpanel listing views -->
   <include package=".analysisprofiles" />
+  <include package=".attachmenttypes" />
   <include package=".departments" />
   <include package=".instrumentlocations" />
   <include package=".interpretationtemplates" />
