+++ src/senaite/core/interfaces/__init__.py
@@ -178,6 +178,16 @@ class IAnalysisProfiles(Interface):
     """
 
 
+class IAttachmentType(Interface):
+    """Marker interface for attachment type
+    """
+
+
+class IAttachmentTypes(Interface):
+    """Marker interface for attachment types setup folder
+    """
+
+
 class ISampleContainers(Interface):
     """Marker interface for sample container setup folder
     """
