+++ src/senaite/core/upgrade/v02_06_000.py
@@ -66,6 +66,8 @@ REMOVE_AT_TYPES = [
     "SamplePreservations",
     "SampleTemplate",
     "SampleTemplates",
+    "AttachmentType",
+    "AttachmentTypes",
 ]
 
 CONTENT_ACTIONS = [
@@ -1057,6 +1059,53 @@ def migrate_samplepoint_to_dx(src, destination=None):
     migrator.copy_id(src, target)
 
     return target
+
+
+@upgradestep(product, version)
+def migrate_attachmenttypes_to_dx(tool):
+    """Converts existing attachment types to Dexterity
+    """
+    logger.info("Convert AttachmentTypes to Dexterity ...")
+
+    # ensure old AT types are flushed first
+    remove_at_portal_types(tool)
+
+    # run required import steps
+    tool.runImportStepFromProfile(profile, "typeinfo")
+    tool.runImportStepFromProfile(profile, "workflow")
+
+    # get the old container
+    origin = api.get_setup().get("bika_attachmnenttypes")
+    if not origin:
+        # old container is already gone
+        return
+
+    # get the destination container
+    destination = get_setup_folder("attachmenttypes")
+
+    # un-catalog the old container
+    uncatalog_object(origin)
+
+    # Mapping from schema field name to a tuple of
+    # (accessor, target field name, default value)
+    schema_mapping = {
+        "title": ("Title", "title", ""),
+        "description": ("Description", "description", ""),
+    }
+
+    # migrate the contents from the old AT container to the new one
+    migrate_to_dx("AttachmentType", origin, destination, schema_mapping)
+
+    # copy snapshots for the container
+    copy_snapshots(origin, destination)
+
+    # remove old AT folder
+    if len(origin) == 0:
+        delete_object(origin)
+    else:
+        logger.warn("Cannot remove {}. Is not empty".format(origin))
+
+    logger.info("Convert AttachmentTypes to Dexterity [DONE]")
 
 
 def update_content_actions(tool):
