+++ src/senaite/core/upgrade/v02_06_000.zcml
@@ -4,6 +4,14 @@
     i18n_domain="senaite.core">
 
   <genericsetup:upgradeStep
+      title="SENAITE.CORE 2.6.0: Migrate ContainerTypes to DX"
+      description="Migrate AttachmentTypes to Dexterity"
+      source="2621"
+      destination="2622"
+      handler=".v02_06_000.migrate_attachmenttypes_to_dx"
+      profile="senaite.core:default"/>
+
+  <genericsetup:upgradeStep
       title="SENAITE.CORE 2.6.0: Migrate SamplePoints to DX"
       description="Migrate SamplePoints to Dexterity"
       source="2619"
